+++ src/index.tsx
@@ -48,6 +48,7 @@ export * from './components/TextArea'
 export * from './components/TextInput'
 export * from './components/TimezonePicker'
 export { getOptions as getTimezoneOptions } from './components/TimezonePicker/utils'
+export * from './components/Toggle'
 export * from './components/Tooltip'
 export { TabPanel, Tabs } from 'react-tabs'